New operating mode: SaveOnly
I found myself using an empty jobfile with yeet mode often to take
advantage of the fact that it saves the config even if there are no changes.

This new mode will only save the config and does not require any jobfile,
eliminating the need to keep empty jobfiles and reducing the risk of accidentally
sending a non-empty one and applying a config you didn't mean to.