Allow overriding the output directory.
This moves p_config to a module level global and accepts a -o CLI
argument to override the default output folder. This will allow a
script to reuse the same output directory for like but separate
jobs.

With this change, `p_config` was promoted from a function argument to a
module level global.